Add base slurm job script to check functionality
